파일 해쉬화 닫기 버튼 기능 연결
