+++ public/utils/channel-common.js
@@ -22,7 +22,6 @@ import { SwiperFunc } from "./parts/channel/swiper.js";
 import { fileHash } from "./parts/channel/fileHash.js";
 import { contentFunc } from "./parts/channel/content.js";
 import { pdfFunc } from "./parts/channel/pdfShare.js";
-import { identifireFunc } from "./parts/channel/identifire.js";
 
 $(async () => {
     mobileDisplayCtr();
@@ -45,7 +44,6 @@ $(async () => {
     fileHash();
     contentFunc();
     pdfFunc();
-    identifireFunc();
 });
 
 // FileDelivery Activator Functions
