+++ views/popup/identifier.ejs
@@ -10,7 +10,7 @@
 <div class="identifier-container" hidden="true">
     <div class="modal-header">
         <h5 class="modal-title" id="modal-label">File Indentity Verification</h5>
-        <button class="identifire-section-close">&times;</button>
+        <button id="closeBtn" class="identifire-section-close">&times;</button>
     </div>
     <div class="modal-body identifier-body">
         <section class="form-signin">
